mcp-servers/x402-pay/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
eth-account>=0.10.0
eciespy>=0.4.0
python-dotenv>=1.0.0
web3>=6.14.0
